requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25         # El ORM más potente
alembic>=1.13.1            # Para migraciones de BD (versionado de tablas)
pymysql>=1.1.0             # Driver MySQL
asyncmy>=0.2.9             # Driver MySQL Asíncrono (Vital para FastAPI)
pydantic-settings>=2.1.0   # Gestión de variables de entorno (.env)
python-multipart>=0.0.9    # Subida de archivos (Excel)
httpx>=0.26.0              # Cliente HTTP asíncrono (Para hablar con MikroTik)
passlib[bcrypt]>=1.7.4     # Seguridad contraseñas
python-jose[cryptography]  # Tokens JWT
pandas>=2.2.0              # Para importar Excel
openpyxl>=3.1.2            # Dependencia de Pandas